Refactor Prompt Templates into Generator Component Structure
- Moved prompt templates from `api/prompts.py` to `api/components/generator/templates/templates.py` to enhance organization and maintainability.
- Updated import statements across the codebase to reflect the new structure, ensuring all existing functionality is preserved.
- Added exports in `api/components/generator/templates/__init__.py` for easier access to templates.
- Completed Phase 7.3 of the API restructure, improving the modularity of the prompt management system.
