+++ memory-bank/tasks/TASK018-phase-7-3-prompts.md
@@ -1,6 +1,6 @@
 # [TASK018] - Phase 7.3: Prompts (From prompts.py)
 
-**Status:** 🔴 Not Started (0%)  
+**Status:** 🟢 Completed (100%)  
 **Priority:** 🟡 Medium  
 **Assignee:** AI Assistant  
 **Created:** 2025-08-27  
@@ -24,16 +24,16 @@ This is a relatively straightforward move but important for keeping the generato
 
 ## Progress Tracking
 
-**Overall Status:** Not Started - 0%
+**Overall Status:** Completed - 100%
 
 ### Subtasks
 | ID | Description | Status | Updated | Notes |
 |----|-------------|--------|---------|-------|
-| 7.3.1 | Move to `components/generator/templates/templates.py` | Not Started | 2025-08-27 | Template relocation |
-| 7.3.2 | Organize prompt management | Not Started | 2025-08-27 | Improve organization |
-| 7.3.3 | Preserve existing prompt functionality | Not Started | 2025-08-27 | Maintain all prompts |
-| 7.3.4 | Update imports and references | Not Started | 2025-08-27 | Fix template references |
-| 7.3.5 | Test prompt template functionality | Not Started | 2025-08-27 | Validate template usage |
+| 7.3.1 | Move to `components/generator/templates/templates.py` | Completed | 2025-08-28 | Moved prompts and created module |
+| 7.3.2 | Organize prompt management | Completed | 2025-08-28 | Re-exported via `templates/__init__.py` |
+| 7.3.3 | Preserve existing prompt functionality | Completed | 2025-08-28 | Templates unchanged, centralized |
+| 7.3.4 | Update imports and references | Completed | 2025-08-28 | Updated all references to new path |
+| 7.3.5 | Test prompt template functionality | Completed | 2025-08-28 | Imports and dependent tests pass; unrelated tests failing |
 
 ## Progress Log
 ### 2025-08-27
@@ -41,17 +41,24 @@ This is a relatively straightforward move but important for keeping the generato
 - Set up subtasks for prompt template relocation
 - Ready for implementation to begin
 
+### 2025-08-28
+- Implemented prompt relocation to `api/components/generator/templates/templates.py`
+- Added exports in `api/components/generator/templates/__init__.py`
+- Updated imports in `api/pipelines/rag/steps/response_generation.py`, `api/rag.py`, `api/pipelines/chat/steps.py`, and `api/simple_chat.py`
+- Removed `api/prompts.py`
+- Ran test suite; prompt-related functionality works. Remaining failures are unrelated to prompts (retriever tests and an API fixture).
+
 ## Dependencies
 - TASK004: Generator components should be available
 - TASK002: Directory structure must be created
 
 ## Success Criteria
-- [ ] Prompt templates moved to generator component structure
-- [ ] Prompt management organized properly
-- [ ] All existing prompt functionality preserved
-- [ ] Import statements updated correctly
-- [ ] Prompt templates working in new location
-- [ ] Template organization improved
+- [x] Prompt templates moved to generator component structure
+- [x] Prompt management organized properly
+- [x] All existing prompt functionality preserved
+- [x] Import statements updated correctly
+- [x] Prompt templates working in new location
+- [x] Template organization improved
 
 ## Risks
 - **Low Risk**: Prompt relocation is typically straightforward
